Extract `useGetCartEntry` helper hook
