+++ apps/store/src/components/CartInventory/CartInventory.helpers.tsx
@@ -1,36 +1,17 @@
 import { useTranslation } from 'next-i18next'
+import { useCallback } from 'react'
 import {
   CampaignDiscount,
   CampaignDiscountType,
   ExternalInsuranceCancellationOption,
+  RedeemedCampaign,
 } from '@/services/apollo/generated'
 import { ShopSession } from '@/services/shopSession/ShopSession.types'
 import { convertToDate } from '@/utils/date'
 import { Money } from '@/utils/formatter'
+import { useGetDiscountExplanation } from '@/utils/useDiscountExplanation'
 import { useFormatter } from '@/utils/useFormatter'
-import { CartEntry } from './CartInventory.types'
-
-export const useGetDiscountDurationExplanation = () => {
-  const { t } = useTranslation('cart')
-  const formatter = useFormatter()
-
-  return (discount: CampaignDiscount, total: Money) => {
-    switch (discount.type) {
-      case CampaignDiscountType.FreeMonths:
-      case CampaignDiscountType.MonthlyPercentage:
-        return t('DISCOUNT_DURATION_EXPLANATION', {
-          count: discount.months,
-          monthlyPrice: formatter.monthlyPrice(total),
-          // Avoid double escaping / in price.  Safe since we're not using dangerouslySetInnerHtml
-          interpolation: { escapeValue: false },
-        })
-      case CampaignDiscountType.MonthlyCost:
-      case CampaignDiscountType.IndefinitePercentage:
-      default:
-        return ''
-    }
-  }
-}
+import { CartCampaign, CartEntry } from './CartInventory.types'
 
 export const getTotal = (shopSession: ShopSession) => {
   if (!shopSession.cart.redeemedCampaign) return shopSession.cart.cost.net
@@ -78,4 +59,45 @@ const getTierLevelDisplayName = (item: ShopSession['cart']['entries'][number]) =
   return item.variant.displayName !== item.variant.product.displayNameFull
     ? item.variant.displayName
     : undefined
+}
+
+type GetCartCampaignFunction = (total: Money, campaign: RedeemedCampaign) => CartCampaign
+
+export const useGetCartCampaign = (): GetCartCampaignFunction => {
+  const getDiscountExplanation = useGetDiscountExplanation()
+  const getDiscountDurationExplanation = useGetDiscountDurationExplanation()
+
+  return useCallback(
+    (total, campaign) => {
+      return {
+        id: campaign.id,
+        code: campaign.code,
+        discountExplanation: getDiscountExplanation(campaign.discount),
+        discountDurationExplanation: getDiscountDurationExplanation(campaign.discount, total),
+      }
+    },
+    [getDiscountExplanation, getDiscountDurationExplanation],
+  )
+}
+
+const useGetDiscountDurationExplanation = () => {
+  const { t } = useTranslation('cart')
+  const formatter = useFormatter()
+
+  return (discount: CampaignDiscount, total: Money) => {
+    switch (discount.type) {
+      case CampaignDiscountType.FreeMonths:
+      case CampaignDiscountType.MonthlyPercentage:
+        return t('DISCOUNT_DURATION_EXPLANATION', {
+          count: discount.months,
+          monthlyPrice: formatter.monthlyPrice(total),
+          // Avoid double escaping / in price.  Safe since we're not using dangerouslySetInnerHtml
+          interpolation: { escapeValue: false },
+        })
+      case CampaignDiscountType.MonthlyCost:
+      case CampaignDiscountType.IndefinitePercentage:
+      default:
+        return ''
+    }
+  }
 }
